Fixed Calendar
Added the appropriate changes in the change log, this was a error due to
the new google reformating coming from a product release at IO I would
expect. This was checked cross 2 servers for compatibility and error
analysis and I can conclude it was directly the hosts fault and not the
developers.